Second commit
First steps into the chat configuration.
Need changes in the dashboard (the list of chats need to be under de toolspace above).
The chat creation box need the members configuration (need to have a checklist so we can list the members of the group).
The chat page need changes and needs to store the chat (maybe a blackout 1 in 1 day).